pyproject.toml
[project]
name = "webasto-next-modbus"
version = "0.1.0"
description = "Home Assistant integration for Webasto Next / Ampure Unite wallboxes via Modbus TCP"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-homeassistant-custom-component>=0.13",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.8"
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "C4", "TID"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
